split into `GeometryBasicsCore`
