+++ GeometryBasicsCore/Project.toml
@@ -1,0 +1,18 @@
+name = "GeometryBasicsCore"
+uuid = "609001ca-df33-43ad-af5e-98eca4304ed0"
+authors = ["SimonDanisch <sdanisch@gmail.com>"]
+version = "1.0.0"
+
+[deps]
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+
+[compat]
+StaticArrays = "0.12, 1.0"
+julia = "1.6"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
