+++ GeometryBasicsCore/src/geometry_primitives.jl
@@ -125,3 +125,19 @@ function normals(vertices::AbstractVector{<:AbstractPoint{3,T}}, faces::Abstract
     normals_result .= normalize.(normals_result)
     return normals_result
 end
+
+const FaceMesh{Dim,T,Element} = Mesh{Dim,T,Element,<:FaceView{Element}}
+
+coordinates(mesh::FaceMesh) = coordinates(getfield(mesh, :simplices))
+faces(mesh::FaceMesh) = faces(getfield(mesh, :simplices))
+
+function texturecoordinates(mesh::AbstractMesh)
+    hasproperty(mesh, :uv) && return mesh.uv
+    hasproperty(mesh, :uvw) && return mesh.uvw
+    return nothing
+end
+
+function normals(mesh::AbstractMesh)
+    hasproperty(mesh, :normals) && return mesh.normals
+    return nothing
+end
