+++ GeometryBasicsCore/src/offsetintegers.jl
