+++ GeometryBasicsCore/src/viewtypes.jl
@@ -87,21 +87,6 @@ end
     return reinterpret(Face(P, T), TupleView{N,skip}(points))
 end
 
-@inline function connect(points::AbstractMatrix{T},
-                         P::Type{<:AbstractPoint{N}}) where {T <: Real,N}
-    return if size(points, 1) === N
-        return reinterpret(Point{N,T}, points)
-    elseif size(points, 2) === N
-        seglen = size(points, 1)
-        columns = ntuple(N) do i
-            return view(points, ((i - 1) * seglen + 1):(i * seglen))
-        end
-        return StructArray{Point{N,T}}(columns)
-    else
-        error("Dim 1 or 2 must be equal to the point dimension!")
-    end
-end
-
 """
     FaceView{Element, Point, Face, P, F}
 
@@ -134,8 +119,6 @@ end
 function Base.propertynames(faceview::FaceView)
     return propertynames(getfield(faceview, :elements))
 end
-
-Tables.schema(faceview::FaceView) = Tables.schema(getfield(faceview, :elements))
 
 Base.size(faceview::FaceView) = size(getfield(faceview, :faces))
